wip(Code): Rename module columnviews to views
