+++ MiAZ/data/docs/VERSION
@@ -1,1 +1,1 @@
-0.0.1+build.5756
+0.0.1+build.5757
