+++ MiAZ/frontend/desktop/widgets/views.py
